backend/products/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0

sqlalchemy==2.0.30
asyncpg==0.29.0

python-dotenv==1.0.1

pydantic==2.7.1
pydantic-settings==2.2.1

aiokafka==0.10.0